+++ src/main/java/com/wow/delivery/dto/user/UserSignupDTO.java
@@ -1,5 +1,6 @@
 package com.wow.delivery.dto.user;
 
+import com.wow.delivery.dto.common.PasswordEncodingDTO;
 import com.wow.delivery.entity.User;
 import com.wow.delivery.util.PasswordEncoder;
 import jakarta.validation.constraints.NotBlank;
